refactor(runner): deepen adapter declarations
