+++ packages/runner/src/adapters.ts
@@ -1,48 +1,54 @@
-import { execFile, spawn, type ChildProcess } from "node:child_process";
-import { createHash } from "node:crypto";
-import { join } from "node:path";
-import { fileURLToPath } from "node:url";
-
 import { stepRole } from "@anneal/db";
 
-import type { ClaimedTask, FailureClass } from "./api.js";
-import type { InFlightTool } from "./budget.js";
+import type { ClaimedTask } from "./api.js";
 import type { RunnerConfig, RunnerKind } from "./config.js";
-import { isTransientNetworkError } from "./network-retry.js";
 import { manifestLines, toolsFor, type SessionToolTransport } from "./session-tool-contract.js";
 import type { AgentScratch } from "./workspace.js";
-import { createClaudeAdapter, claudeArgs, claudeChildEnvironment, provisionClaudeSessionConfig } from "./adapters/claude.js";
-import {
-  CODEX_STARTER_MODEL, codexArgs, codexChildEnvironment, codexNativeSubagentProfile, codexPlatformBaselinePath, createCodexAdapter,
-  provisionCodexSessionConfig,
-} from "./adapters/codex.js";
+import { claudeDeclaration, claudePlatformSettingsPath } from "./adapters/claude.js";
+import { codexDeclaration, codexPlatformBaselinePath } from "./adapters/codex.js";
 import { workspaceEnvironment } from "./adapters/environment.js";
-import { createPiAdapter, piArgs, piChildEnvironment, provisionPiSessionConfig } from "./adapters/pi.js";
+import { piDeclaration, piExtensionPath } from "./adapters/pi.js";
+import {
+  createCliAdapter,
+  launchAdapterArgv,
+  mcpServerPath,
+  nodeBinaryPath,
+  type AdapterDeclaration,
+  type CliAdapter,
+  type ResumeSpec,
+  type RunSpec,
+} from "./adapters/runtime.js";
 import type { SessionConfigOptions } from "./adapters/session-config.js";
+
+export * from "./adapters/runtime.js";
+export { claudePlatformSettingsPath } from "./adapters/claude.js";
+export { piExtensionPath } from "./adapters/pi.js";
 
 export const ADAPTER_VERSION = "2.1.0";
 
-export type ToolKey = "BASH" | "READ" | "WRITE" | "EDIT" | "GLOB" | "GREP" | "WEB_FETCH" | "WEB_SEARCH";
-export const TOOL_ORDER: ToolKey[] = ["BASH", "READ", "WRITE", "EDIT", "GLOB", "GREP", "WEB_FETCH", "WEB_SEARCH"];
-export const CLAUDE_TOOL_NAMES: Record<ToolKey, string> = {
-  BASH: "Bash", READ: "Read", WRITE: "Write", EDIT: "Edit",
-  GLOB: "Glob", GREP: "Grep", WEB_FETCH: "WebFetch", WEB_SEARCH: "WebSearch",
-};
-export const PI_TOOL_NAMES: Partial<Record<ToolKey, string>> = {
-  BASH: "bash", READ: "read", WRITE: "write", EDIT: "edit",
-};
+const ADAPTER_DECLARATIONS: Readonly<Record<RunnerKind, AdapterDeclaration>> = Object.freeze({
+  CLAUDE: claudeDeclaration,
+  CODEX: codexDeclaration,
+  PI: piDeclaration,
+});
 
-// The seat manual tells the agent to use the Anneal tools; the manifest has to
-// name them, or the agent has no way to know what it was actually granted.
+const COMMON_PROTECTED_SECRET_ENVIRONMENT = [
+  "GIT_CONFIG_GLOBAL", "AGENTOS_GATE_SERVER",
+  "HTTP_PROXY", "HTTPS_PROXY", "NO_PROXY", "http_proxy", "https_proxy", "no_proxy",
+] as const;
+
+const PROTECTED_SECRET_ENVIRONMENT = new Set([
+  ...COMMON_PROTECTED_SECRET_ENVIRONMENT,
+  ...Object.values(ADAPTER_DECLARATIONS).flatMap((declaration) => declaration.protectedEnvironmentVariables),
+]);
+
 const toolManifest = (claim: ClaimedTask): string[] => [
   "",
   RUNNER_DEFINITIONS[claim.runner].toolIntroduction,
   ...manifestLines(RUNNER_DEFINITIONS[claim.runner].toolTransport),
 ];
 
-export const buildPrompt = (claim: ClaimedTask): string => {
-  const subagents = codexNativeSubagentProfile(claim.run, claim.runner);
-  return [
+export const buildPrompt = (claim: ClaimedTask): string => [
   claim.agent.foundationalPrompt,
   "",
   `Role (${claim.agent.name}): ${claim.agent.rolePrompt}`,
@@ -68,17 +74,7 @@ export const buildPrompt = (claim: ClaimedTask): string => {
     `- implementationBaseSha: ${claim.run.implementationBaseSha}`,
     `- implementationHeadSha: ${claim.run.implementationHeadSha}`,
   ] : []),
-  ...(subagents ? [
-    "",
-    "Platform-pinned native implementation subagents:",
-    `- model: ${subagents.model}`,
-    `- reasoning effort: ${subagents.effort}`,
-    `- maximum concurrent child threads: ${subagents.maxConcurrent} (root excluded)`,
-    "- multi_agent_v2 is enabled by the runner. Spawn, message, wait for, and close native children through the session collaboration tools; do not launch nested Codex CLI processes.",
-    "- The runner enforces the same child model and concurrency snapshot on fresh starts and resumes. Do not select or escalate a child model.",
-    "- Implementation proof is limited to each assignment's focused tests, one affected-workspace compile or typecheck after integration, and tests for seams crossed by multiple assignments.",
-    "- Do not run repository-wide suites or the repository Merge Gate in Implementation; the later Regression step owns the formal Gate.",
-  ] : []),
+  ...ADAPTER_DECLARATIONS[claim.runner].promptSections(claim),
   "",
   `Task: ${claim.task.name}`,
   claim.task.description,
@@ -131,8 +127,7 @@ export const buildPrompt = (claim: ClaimedTask): string => {
     ]),
     `- Repair task output (${claim.regressionRepairHandoff.repair.outputKind}):\n${claim.regressionRepairHandoff.repair.outputBody}`,
   ] : []),
-  ].join("\n");
-};
+].join("\n");
 
 export const buildChildEnvironment = (
   config: Pick<RunnerConfig, "path" | "home" | "apiUrl" | "runAsPrefix">
@@ -147,28 +142,9 @@ export const buildChildEnvironment = (
   if (regressionStep && !claim.task.chainId) {
     throw new Error("regression-verification task is missing its platform chain id");
   }
-  // These names are runner-owned. In particular, a task secret must never be
-  // able to reintroduce the host Codex home or the CLAUDE_CONFIG_DIR path that
-  // this chain deliberately does not use.
-  const {
-    CLAUDE_CONFIG_DIR: _claudeConfigDir,
-    CODEX_HOME: _codexHome,
-    GIT_CONFIG_GLOBAL: _gitConfigGlobal,
-    PI_CODING_AGENT_DIR: _piCodingAgentDir,
-    PI_CODING_AGENT_SESSION_DIR: _piCodingAgentSessionDir,
-    AGENTOS_CODEX_SERVICE_TIER: _codexServiceTier,
-    AGENTOS_PI_EXPECTS_OPENAI_CODEX: _piExpectsOpenAICodex,
-    AGENTOS_GATE_SERVER: _gateServer,
-    HTTP_PROXY: _httpProxy,
-    HTTPS_PROXY: _httpsProxy,
-    NO_PROXY: _noProxy,
-    http_proxy: _httpProxyLower,
-    https_proxy: _httpsProxyLower,
-    no_proxy: _noProxyLower,
-    ...unfilteredTaskSecrets
-  } = claim.secrets;
-  const taskSecrets = Object.fromEntries(Object.entries(unfilteredTaskSecrets).filter(([name]) =>
-    !name.startsWith("GIT_CONFIG_")
+  const taskSecrets = Object.fromEntries(Object.entries(claim.secrets).filter(([name]) =>
+    !PROTECTED_SECRET_ENVIRONMENT.has(name)
+    && !name.startsWith("GIT_CONFIG_")
     && !["GIT_AUTHOR_NAME", "GIT_AUTHOR_EMAIL", "GIT_COMMITTER_NAME", "GIT_COMMITTER_EMAIL"].includes(name)));
   return {
     ...taskSecrets,
@@ -187,102 +163,20 @@ export const buildChildEnvironment = (
       AGENTOS_CHAIN_ID: claim.task.chainId!,
       AGENTOS_PULL_REQUEST_BASE: claim.run.pullRequestBase,
     } : {}),
-    AGENTOS_CODEX_SERVICE_TIER: claim.run.codexServiceTier.toLowerCase(),
     ...RUNNER_DEFINITIONS[claim.runner].childEnvironment(claim, scratch),
-    // Last on purpose, so no task secret can point a session back at the
-    // production roots. See provisionAgentScratch for why this containment has
-    // to live in the runner rather than in the run's checkout.
     RUNNER_WORKSPACE_ROOT: scratch.workspaceRoot,
     CONTROL_PLANE_STATE_DIR: scratch.stateDir,
   };
 };
 
-const isolationVariables = [
-  "RUNNER_WORKSPACE_ROOT", "CONTROL_PLANE_STATE_DIR", "HOME", "GIT_CONFIG_GLOBAL", "CODEX_HOME", "PI_CODING_AGENT_DIR",
-  "AGENTOS_CODEX_SERVICE_TIER", "AGENTOS_PI_EXPECTS_OPENAI_CODEX", "AGENTOS_GATE_SERVER",
-  "AGENTOS_RUN_ID", "GIT_CONFIG_COUNT", "GIT_CONFIG_KEY_0", "GIT_CONFIG_VALUE_0",
-] as const;
-
-/**
- * The command line for a session, re-asserting the isolation variables on the
- * far side of RUNNER_RUN_AS_PREFIX.
- *
- * The prefix is an arbitrary launcher, and a launcher is free to scrub the
- * environment it was handed. #117's shipped example is `sudo -u _agentos1 -E
- * --`, where everything the session inherits rests on that `-E`: without it
- * sudo's env_reset strips the lot. Putting the two roots only in the
- * environment we pass to the *launcher* is therefore not a guarantee.
- *
- * Losing them is also the one failure that is both silent and catastrophic: an
- * old base falls back to the production default and sweeps it, which is #125.
- * Everything else a scrubbing launcher drops — PATH and
- * AGENTOS_SESSION_TOKEN — fails loudly and immediately instead. HOME is a
- * containment variable because Codex discovers ~/.agents/skills independently
- * of CODEX_HOME; losing it could silently expose the launched account's home.
- * GIT_CONFIG_GLOBAL travels with it so that relocation does not silently
- * replace the configured runner identity or drop credential helpers.
- *
- * So the containment variables and the non-secret, Run-snapshotted Codex
- * profile are set again by `/usr/bin/env`, which runs after
- * the launcher and immediately before the CLI. That also makes the contract
- * fail-closed: a launcher that will not exec `/usr/bin/env` cannot start the
- * session at all, rather than starting it pointed at production.
- *
- * Proxy values deliberately stay in the inherited environment instead of this
- * argv: proxy URLs may contain credentials, and argv is readable through `ps`.
- * A configured RUNNER_RUN_AS_PREFIX therefore has to preserve the environment
- * it receives (the shipped sudo prefix uses `-E`). This is the same channel the
- * runner's Git and workspace commands use, so a prefix that scrubs it cannot
- * provide the platform-wide proxy contract.
- */
+/** Re-assert containment and provider-owned policy after a scrubbing launcher. */
 export const launchArgv = (
   config: Pick<RunnerConfig, "runAsPrefix" | "binaries">,
   runner: RunnerKind,
   args: string[],
   env: NodeJS.ProcessEnv,
-): { executable: string; args: string[] } => {
-  const binary = config.binaries[runner];
-  if (config.runAsPrefix.length === 0) return { executable: binary, args };
-  const assignments = isolationVariables.flatMap((name) => env[name] !== undefined ? [`${name}=${env[name]}`] : []);
-  return {
-    executable: config.runAsPrefix[0]!,
-    args: [
-      ...config.runAsPrefix.slice(1),
-      ...(assignments.length > 0 ? ["/usr/bin/env", ...assignments] : []),
-      binary,
-      ...args,
-    ],
-  };
-};
+): { executable: string; args: string[] } => launchAdapterArgv(config, ADAPTER_DECLARATIONS[runner], args, env);
 
-// Resolved from the package root so it works from dist/ (launchd) and src/ (tsx).
-const packageRoot = fileURLToPath(new URL("..", import.meta.url));
-
-export const mcpServerPath = (): string => process.env.RUNNER_MCP_SERVER_PATH ?? join(packageRoot, "dist", "mcp-server.js");
-export const piExtensionPath = (): string => process.env.RUNNER_PI_EXTENSION_PATH ?? join(packageRoot, "assets", "pi-agentos-extension.ts");
-export const claudePlatformSettingsPath = (): string =>
-  process.env.RUNNER_CLAUDE_SETTINGS_PATH ?? join(packageRoot, "assets", "claude-platform-settings.json");
-
-/**
- * The interpreter the CLI is told to run the MCP server with.
- *
- * The default is this daemon's own `process.execPath`, which is a resolved real
- * path — typically a Homebrew Cellar or version-manager directory, not the
- * symlink in `RUNNER_PATH`. Under a run-as prefix the CLI is a different account
- * that may not be able to traverse it, and the failure surfaces as an MCP
- * startup error inside the session rather than as a deployment error. The
- * override exists so a deployment can name an interpreter both principals can
- * execute, and `runtimeDescriptor()` publishes whichever one is in force so it
- * can be checked from outside the process.
- */
-export const nodeBinaryPath = (): string => process.env.RUNNER_NODE_BINARY ?? process.execPath;
-
-/**
- * One machine-readable line naming every path the agent's session depends on and
- * that a different OS principal has to be able to reach. Printed at startup so
- * `scripts/os-isolation/verify.sh` can test the paths the daemon actually uses
- * instead of whatever the operator's own shell happens to resolve.
- */
 export const runtimeDescriptor = (runnerId: string, runAsPrefix: string[]): string => JSON.stringify({
   runtime: "agentos-runner",
   runnerId,
@@ -295,553 +189,6 @@ export const runtimeDescriptor = (runnerId: string, runAsPrefix: string[]): stri
   runAsPrefix: runAsPrefix.join(" "),
 });
 
-/**
- * The Anneal MCP server the agent's CLI session spawns for itself. It carries
- * no credentials: they reach the server through the inherited child environment,
- * so nothing secret ends up in a command line other processes can read.
- */
-export const mcpServerArgs = (credentialsPath: string): string[] => [mcpServerPath(), "--credentials", credentialsPath];
-
-export const mcpConfig = (credentialsPath: string): { mcpServers: Record<string, { type: string; command: string; args: string[] }> } => ({
-  mcpServers: { agentos: { type: "stdio", command: nodeBinaryPath(), args: mcpServerArgs(credentialsPath) } },
-});
-
-export type AdapterEvent = {
-  source: "RUNNER" | "CLAUDE" | "CODEX" | "PI";
-  type: string;
-  providerEventId?: string | null;
-  toolCallId?: string | null;
-  payload: Record<string, unknown>;
-};
-
-export type SessionEventSink = (event: AdapterEvent) => void;
-
-export type ExitEvidence = {
-  exitCode: number | null;
-  signal: string | null;
-  terminalEventSeen: boolean;
-  terminalSuccess: boolean;
-  terminationReason: string | null;
-  finalOutput: string | null;
-  providerError: string | null;
-  stdout: string;
-  stderr: string;
-};
-
-export type ClassifiedFailure = {
-  failureClass: FailureClass;
-  retryable: boolean;
-  operatorAction?: string;
-};
-
-export type HeartbeatSnapshot = {
-  processAlive: boolean;
-  lastProcessAliveAt: Date;
-  lastProgressEventAt: Date;
-  inFlightTool: InFlightTool | null;
-};
-
-/**
- * A PI session's harvested token and cost totals. Every field is `null` until a
- * message actually reports it — absent is never zero, matching the
- * `SessionUsage` contract these totals are ingested through.
- *
- * `messages` and `reported` are diagnostics carried into the payload: they are
- * what lets an operator tell "the session sent nothing" apart from "the session
- * sent messages and PI reported no usage for them".
- */
-export type PiUsageTotals = {
-  messages: number;
-  reported: number;
-  input: number | null;
-  output: number | null;
-  cacheRead: number | null;
-  cacheWrite: number | null;
-  costNanoUsd: number | null;
-};
-
-export type AdapterState = {
-  runId: string;
-  runner: RunnerKind;
-  startedAt: Date;
-  lastProcessAliveAt: Date;
-  lastProgressEventAt: Date;
-  inFlightTool: InFlightTool | null;
-  providerConversationId: string | null;
-  terminalEventSeen: boolean;
-  terminalSuccess: boolean;
-  terminationReason: string | null;
-  sawError: boolean;
-  providerError: string | null;
-  piTurnCompleted: boolean;
-  piFinalAttemptFailed: boolean;
-  piUsage: PiUsageTotals;
-  finalOutput: string | null;
-  stdout: string;
-  stderr: string;
-};
-
-export type RuntimeHandle = AdapterState & {
-  child: ChildProcess;
-  pid: number | null;
-  exit: Promise<ExitEvidence>;
-};
-
-export type PreflightSpec = {
-  config: RunnerConfig;
-  runner: RunnerKind;
-  model: string | null;
-  env: NodeJS.ProcessEnv;
-};
-
-export type PreflightResult = {
-  ok: boolean;
-  cliVersion: string | null;
-  authMode: string | null;
-  capabilities: Record<string, unknown>;
-  error?: string;
-};
-
-export type RunSpec = {
-  config: RunnerConfig;
-  claim: ClaimedTask;
-  workingDirectory: string;
-  env: NodeJS.ProcessEnv;
-  prompt: string;
-  /** 0600 file the Anneal MCP server reads its session credentials from. */
-  credentialsPath: string;
-};
-
-export type ResumeSpec = RunSpec & { providerConversationId: string; input: string };
-
-export type KillResult = { signal: "SIGTERM" | "SIGKILL" | null; processAlive: boolean };
-
-/**
- * One CLI implementation at the runner seam.
- *
- * `claim.agent.disabledTools` is not a universal capability: Claude enforces
- * all eight current tool keys, PI enforces BASH/READ/WRITE/EDIT, and Codex
- * exposes no supported per-tool deny mechanism. An adapter must keep that
- * limitation explicit in its argv implementation; callers must not infer that
- * a non-empty disabledTools list was enforced merely because start succeeded.
- */
-export interface CliAdapter {
-  preflight(spec: PreflightSpec): Promise<PreflightResult>;
-  start(spec: RunSpec, sink: SessionEventSink): Promise<RuntimeHandle>;
-  resume(spec: ResumeSpec, sink: SessionEventSink): Promise<RuntimeHandle>;
-  kill(handle: RuntimeHandle, reason: string): Promise<KillResult>;
-  heartbeat(handle: RuntimeHandle): Promise<HeartbeatSnapshot>;
-  classifyError(evidence: ExitEvidence): ClassifiedFailure;
-}
-
-export type AdapterEventParser = (
-  state: AdapterState,
-  event: Record<string, unknown>,
-  sink: SessionEventSink,
-) => void;
-
-export type AdapterImplementation = {
-  runner: RunnerKind;
-  args(spec: RunSpec, resume?: ResumeSpec): string[];
-  parseEvent: AdapterEventParser;
-};
-
-const emptyPiUsage = (): PiUsageTotals =>
-  ({ messages: 0, reported: 0, input: null, output: null, cacheRead: null, cacheWrite: null, costNanoUsd: null });
-
-export const createAdapterState = (
-  runner: RunnerKind,
-  runId: string,
-  startedAt = new Date(),
-): AdapterState => ({
-  runId,
-  runner,
-  startedAt,
-  lastProcessAliveAt: startedAt,
-  lastProgressEventAt: startedAt,
-  inFlightTool: null,
-  providerConversationId: null,
-  terminalEventSeen: false,
-  terminalSuccess: false,
-  terminationReason: null,
-  sawError: false,
-  providerError: null,
-  piTurnCompleted: false,
-  piFinalAttemptFailed: false,
-  piUsage: emptyPiUsage(),
-  finalOutput: null,
-  stdout: "",
-  stderr: "",
-});
-
-const cap = (value: string, limit = 1_000_000): string => value.length <= limit ? value : value.slice(value.length - limit);
-
-const sourceFor = (runner: RunnerKind): AdapterEvent["source"] => runner;
-
-export const asRecord = (value: unknown): Record<string, unknown> | null =>
-  typeof value === "object" && value !== null && !Array.isArray(value) ? value as Record<string, unknown> : null;
-
-export const stringField = (object: Record<string, unknown> | null, field: string): string | null =>
-  typeof object?.[field] === "string" ? object[field] as string : null;
-
-export const eventErrorMessage = (event: Record<string, unknown>): string | null =>
-  stringField(event, "message")
-  ?? stringField(event, "error")
-  ?? stringField(asRecord(event.error), "message");
-
-export const emitAdapterEvent = (state: AdapterState, sink: SessionEventSink, type: string, payload: Record<string, unknown>, toolCallId?: string | null): void => {
-  state.lastProgressEventAt = new Date();
-  sink({ source: sourceFor(state.runner), type, payload, ...(toolCallId !== undefined ? { toolCallId } : {}) });
-};
-
-export const markInFlightToolProgress = (state: AdapterState): void => {
-  if (state.inFlightTool) state.inFlightTool.lastProgressAt = new Date();
-};
-
-export const processProviderEvent = (
-  state: AdapterState,
-  event: Record<string, unknown>,
-  sink: SessionEventSink,
-  parseEvent: AdapterEventParser,
-): void => {
-  sink({ source: sourceFor(state.runner), type: "PROVIDER_RAW", payload: event });
-  parseEvent(state, event, sink);
-};
-
-const processLine = (
-  state: AdapterState,
-  line: string,
-  sink: SessionEventSink,
-  parseEvent: AdapterEventParser,
-): void => {
-  if (!line.trim()) return;
-  let event: Record<string, unknown>;
-  try {
-    const parsed = JSON.parse(line) as unknown;
-    event = asRecord(parsed) ?? { value: parsed };
-  } catch {
-    state.sawError = true;
-    emitAdapterEvent(state, sink, "ADAPTER_ERROR", { error: "invalid-json", line });
-    return;
-  }
-  processProviderEvent(state, event, sink, parseEvent);
-};
-
-// Run.model carries an optional reasoning-effort suffix: "<model>[:<effort>]".
-export const modelSpec = (raw: string): { model: string; effort: string | null } => {
-  const at = raw.lastIndexOf(":");
-  return at > 0 ? { model: raw.slice(0, at), effort: raw.slice(at + 1) } : { model: raw, effort: null };
-};
-
-/**
- * What the session is actually asked to do: a fresh prompt, or the resume input.
- *
- * It never appears in argv. A claim carries the chain's persisted prior outputs
- * verbatim, each capped at 500k by the write endpoint, so a nine-step chain's
- * cumulative prompt is measured in megabytes — past Linux's 128 KiB per-argument
- * MAX_ARG_STRLEN and past macOS's ~1 MiB total argument block. As an argv
- * element that is E2BIG: `spawn` fails before the provider ever starts, and no
- * amount of retrying fixes it.
- *
- * Every runner reads it from stdin instead, on a documented path of its own:
- * `claude -p` with no prompt argument, `codex exec -` and `codex exec resume
- * <id> -` (`-` means "read the instructions from stdin"), and pi, which
- * prepends piped stdin to the initial message in every non-RPC mode. Resume ids,
- * MCP wiring and deny flags stay on argv, where they are small and stable.
- *
- * The prompt also stops being visible in every `ps` on the box, which is where
- * a task description and its predecessors' outputs used to sit in the clear.
- */
-export const inputForRunner = (spec: RunSpec, resume?: ResumeSpec): string => resume?.input ?? spec.prompt;
-
-export const argsForRunner = (runner: RunnerKind, spec: RunSpec, resume?: ResumeSpec): string[] =>
-  RUNNER_DEFINITIONS[runner].args(spec, resume);
-
-export const spawnAdapterRuntime = (implementation: AdapterImplementation, spec: RunSpec, sink: SessionEventSink, resume?: ResumeSpec): RuntimeHandle => {
-  const { runner } = implementation;
-  const binary = spec.config.binaries[runner];
-  const args = implementation.args(spec, resume);
-  const input = inputForRunner(spec, resume);
-  const { executable, args: fullArgs } = launchArgv(spec.config, runner, args, spec.env);
-  const startedAt = new Date();
-  const child = spawn(executable, fullArgs, {
-    cwd: spec.workingDirectory,
-    env: spec.env,
-    // stdin is the prompt channel (see inputForRunner); stdout stays the
-    // structured event protocol every parser below reads.
-    detached: true,
-    stdio: ["pipe", "pipe", "pipe"],
-  });
-  const handle: RuntimeHandle = {
-    ...createAdapterState(runner, spec.claim.run.id, startedAt),
-    child,
-    pid: child.pid ?? null,
-    exit: Promise.resolve({} as ExitEvidence),
-  };
-  let buffer = "";
-  child.stdout!.setEncoding("utf8");
-  child.stderr!.setEncoding("utf8");
-  child.stdout!.on("data", (chunk: string) => {
-    handle.stdout = cap(handle.stdout + chunk);
-    buffer += chunk;
-    const lines = buffer.split(/\r?\n/u);
-    buffer = lines.pop() ?? "";
-    for (const line of lines) processLine(handle, line, sink, implementation.parseEvent);
-  });
-  child.stderr!.on("data", (chunk: string) => {
-    handle.stderr = cap(handle.stderr + chunk);
-    sink({ source: sourceFor(runner), type: "STDERR", payload: { text: chunk } });
-  });
-  handle.exit = new Promise<ExitEvidence>((resolvePromise) => {
-    let settled = false;
-    const finish = (exitCode: number | null, signal: string | null): void => {
-      if (settled) return;
-      settled = true;
-      if (buffer.trim()) processLine(handle, buffer, sink, implementation.parseEvent);
-      resolvePromise({
-        exitCode,
-        signal,
-        terminalEventSeen: handle.terminalEventSeen,
-        terminalSuccess: handle.terminalSuccess,
-        terminationReason: handle.terminationReason,
-        finalOutput: handle.finalOutput,
-        providerError: handle.providerError,
-        stdout: handle.stdout,
-        stderr: handle.stderr,
-      });
-    };
-    child.once("error", (error: NodeJS.ErrnoException) => {
-      handle.sawError = true;
-      handle.stderr = cap(`${handle.stderr}\n${error.message}`);
-      finish(error.code === "ENOENT" ? 127 : 1, null);
-    });
-    child.once("close", (code, signal) => finish(code, signal));
-  });
-  // A CLI that dies before it drains a multi-megabyte prompt (missing binary,
-  // failed auth) closes the pipe under us. That EPIPE is a symptom, never the
-  // diagnosis, so it is recorded as its own event and deliberately kept out of
-  // stderr: the CLI's own exit evidence is what classifies the run.
-  child.stdin!.on("error", (error: NodeJS.ErrnoException) => {
-    sink({
-      source: "RUNNER",
-      type: "PROMPT_DELIVERY_FAILED",
-      payload: { message: error.message, code: error.code ?? null },
-    });
-  });
-  child.stdin!.end(input);
-  sink({ source: "RUNNER", type: "PROCESS_STARTED", payload: {
-    pid: handle.pid,
-    binary,
-    // argv no longer carries the prompt, so this stays safe to persist and read.
-    args,
-    promptTransport: "stdin",
-    promptBytes: Buffer.byteLength(input),
-    // The persisted Run value is the hash available at queue time, before
-    // claim-time handoffs and notes have been appended. Hash the bytes actually
-    // delivered to this provider invocation so this event cannot claim a digest
-    // for a different prompt (and resume inputs are covered too).
-    promptHash: promptHashFor(input),
-  } });
-  return handle;
-};
-
-export const capturePreflight = async (config: RunnerConfig, runner: RunnerKind, args: string[], env: NodeJS.ProcessEnv): Promise<{ code: number | null; stdout: string; stderr: string }> =>
-  new Promise((resolvePromise) => {
-    const launch = launchArgv(config, runner, args, env);
-    const child = spawn(launch.executable, launch.args, {
-      env,
-      stdio: ["ignore", "pipe", "pipe"],
-    });
-    let stdout = "";
-    let stderr = "";
-    let settled = false;
-    const finish = (code: number | null, extra = ""): void => {
-      if (settled) return;
-      settled = true;
-      clearTimeout(timer);
-      resolvePromise({ code, stdout, stderr: `${stderr}${extra}` });
-    };
-    const timer = setTimeout(() => {
-      child.kill("SIGKILL");
-      finish(1, "\npreflight timed out after 30 seconds");
-    }, 30_000);
-    child.stdout.setEncoding("utf8");
-    child.stderr.setEncoding("utf8");
-    child.stdout.on("data", (chunk: string) => { stdout += chunk; });
-    child.stderr.on("data", (chunk: string) => { stderr += chunk; });
-    child.once("error", (error: NodeJS.ErrnoException) => finish(error.code === "ENOENT" ? 127 : 1, `\n${error.message}`));
-    child.once("close", (code) => finish(code));
-  });
-
-/**
- * What a failed preflight is allowed to say about itself.
- *
- * A preflight failure used to report the CLI's own stdout and stderr verbatim.
- * That text is persisted as `RunnerBackendState.circuitReason`, returned by
- * `GET /runners`, and rendered — so whatever an unauthenticated `codex login
- * status` happened to print, including an environment variable, a remote URL
- * with a password in it or a home-directory path, became telemetry and then
- * became a screenshot. Nothing about that channel is bounded, so nothing about
- * it can be sanitised; the fix is not to carry it.
- *
- * What is carried instead is a class: a fixed phrase this file authors, one per
- * way a preflight can fail, plus the exit code, which is a number. The CLI's own
- * account of the failure stays where the operator can already read it — their
- * terminal, running the same official command the guidance names.
- */
-export const PREFLIGHT_CLASS = {
-  cliMissing: "cli-missing",
-  cliIncompatible: "cli-incompatible",
-  notAuthenticated: "not-authenticated",
-  unsupportedModel: "unsupported-model",
-} as const;
-
-export const PREFLIGHT_REASONS = {
-  cliMissing: `${PREFLIGHT_CLASS.cliMissing}: the CLI did not answer --version`,
-  cliIncompatible: `${PREFLIGHT_CLASS.cliIncompatible}: the CLI does not expose the required Anneal exec protocol`,
-  notAuthenticated: `${PREFLIGHT_CLASS.notAuthenticated}: the CLI's own login check did not pass`,
-  unsupportedModel: `${PREFLIGHT_CLASS.unsupportedModel}: an explicit provider/model is required`,
-} as const;
-
-export const preflightFailure = (reason: string, code: number | null): string =>
-  code === null ? reason : `${reason} (exit ${code})`;
-
-export const adapterExecutionSucceeded = (evidence: ExitEvidence): boolean =>
-  evidence.exitCode === 0
-  && evidence.signal === null
-  && evidence.terminationReason === null
-  && evidence.terminalEventSeen
-  && evidence.terminalSuccess;
-
-/**
- * The tail of what a run produced: the agent's own final output when it emitted
- * one, otherwise its stdout. Capped at 500k because it is sent on the wire and
- * stored whole (`Run.output`); the *tail* is kept rather than the head because
- * the end of a run is where its account of itself is.
- *
- * One definition, because two call sites in runner.ts send it: the ordinary
- * completion, and the exception path, which reports a run whose agent had
- * already finished when the runner's own plumbing failed.
- */
-export const outputTail = (evidence: ExitEvidence): string | null =>
-  (evidence.finalOutput ?? evidence.stdout).trim().slice(-500_000) || null;
-
-export const failureReasonFromEvidence = (evidence: ExitEvidence): string =>
-  evidence.providerError?.trim()
-  || evidence.stderr.trim()
-  || `CLI exited with code ${evidence.exitCode}`;
-
-export const classifyRuntimeError = (evidence: ExitEvidence): ClassifiedFailure => {
-  const text = evidence.providerError?.trim()
-    ? `${evidence.providerError}\n${evidence.stderr}`
-    : `${evidence.stderr}\n${evidence.stdout}`;
-  // Auth verdicts must never be read out of stdout: stdout is the agent's
-  // working output, and a task that edits auth code contains "401"/"not
-  // logged in" as content (run cmsy26f2s0ibqmpmx8t6gyltg was classified
-  // AUTH_REQUIRED off a literal 401 in the code it was writing). Provider
-  // error and stderr are the only channels the CLI reports auth trouble on.
-  const authEvidence = `${evidence.providerError ?? ""}\n${evidence.stderr}`;
-  if (evidence.terminationReason) return { failureClass: "CANCELLED_OR_TIMED_OUT", retryable: false };
-  if (evidence.exitCode === 127) {
-    return { failureClass: "BINARY_NOT_FOUND", retryable: false, operatorAction: "Install the configured CLI or repair RUNNER_PATH" };
-  }
-  // Auth outranks transient: a providerError that names an auth failure must
-  // not be retried into a lockout just because the same message also mentions
-  // a dropped connection or server_error.
-  // `not-authenticated` is this file's own preflight class, not CLI text: the
-  // preflight stopped forwarding the CLI's words (they are unbounded and end up
-  // in telemetry), so the classifier recognises the class instead.
-  if (new RegExp(`authentication_failed|\\b401\\b|Missing authentication|No API key found|not logged in|${PREFLIGHT_CLASS.notAuthenticated}`, "iu").test(authEvidence)) {
-    return { failureClass: "AUTH_REQUIRED", retryable: false, operatorAction: "Log the runner account into the CLI" };
-  }
-  if (/connection lost|server_error/iu.test(`${evidence.providerError ?? ""}`)) {
-    return { failureClass: "TRANSIENT_PROVIDER", retryable: true };
-  }
-  if (/\b429\b|rate.?limit|usage.?limit|quota/iu.test(text)) return { failureClass: "RATE_LIMITED", retryable: true };
-  if (isTransientNetworkError(text) || /\b5\d\d\b|provider outage/iu.test(text)) {
-    return { failureClass: "TRANSIENT_PROVIDER", retryable: true };
-  }
-  if (/"isError"\s*:\s*true|"command_execution"[\s\S]{0,500}"status"\s*:\s*"failed"/u.test(text)) {
-    return { failureClass: "TOOL_FAILED", retryable: false };
-  }
-  if (evidence.exitCode === 0 && (!evidence.terminalEventSeen || !evidence.terminalSuccess)) {
-    return { failureClass: "PROTOCOL_ERROR", retryable: true, operatorAction: "Check CLI protocol/version drift" };
-  }
-  return { failureClass: "TASK_FAILED", retryable: false };
-};
-
-const ownedRunProcesses = async (runId: string): Promise<number[]> => new Promise((resolvePromise, rejectPromise) => {
-  const args = process.platform === "darwin"
-    ? ["-Eww", "-axo", "pid=,ppid=,pgid=,command="]
-    : ["-axeww", "-o", "pid=,ppid=,pgid=,command="];
-  execFile("ps", args, { maxBuffer: 64 * 1024 * 1024 }, (error, stdout) => {
-    if (error) {
-      rejectPromise(new Error(`Unable to inspect Run-owned processes: ${error.message}`));
-      return;
-    }
-    const marker = ` AGENTOS_RUN_ID=${runId}`;
-    const pids = stdout.split("\n").flatMap((line) => {
-      if (!line.includes(marker)) return [];
-      const match = /^\s*(\d+)\s+/u.exec(line);
-      return match ? [Number.parseInt(match[1]!, 10)] : [];
-    });
-    resolvePromise(pids.filter((pid) => pid !== process.pid));
-  });
-});
-
-const signalProcesses = (pids: number[], signal: NodeJS.Signals): void => {
-  for (const pid of pids) {
-    try {
-      process.kill(pid, signal);
-    } catch (error: unknown) {
-      if ((error as NodeJS.ErrnoException).code !== "ESRCH") throw error;
-    }
-  }
-};
-
-const waitForRunDrain = async (runId: string, timeoutMs: number): Promise<number[]> => {
-  const deadline = Date.now() + timeoutMs;
-  let remaining = await ownedRunProcesses(runId);
-  while (remaining.length > 0 && Date.now() < deadline) {
-    await new Promise<void>((resolvePromise) => setTimeout(resolvePromise, 50));
-    remaining = await ownedRunProcesses(runId);
-  }
-  return remaining;
-};
-
-export const killRuntime = async (handle: RuntimeHandle, reason: string): Promise<KillResult> => {
-  handle.terminationReason = reason;
-  const pid = handle.pid;
-  if (pid && handle.child.exitCode === null && handle.child.signalCode === null) {
-    try { process.kill(-pid, "SIGTERM"); } catch { /* the ownership scan below is authoritative */ }
-  }
-  let remaining = await ownedRunProcesses(handle.runId);
-  signalProcesses(remaining, "SIGTERM");
-  remaining = await waitForRunDrain(handle.runId, 5_000);
-  if (remaining.length === 0) {
-    await handle.exit;
-    return { signal: pid ? "SIGTERM" : null, processAlive: false };
-  }
-  signalProcesses(remaining, "SIGKILL");
-  remaining = await waitForRunDrain(handle.runId, 5_000);
-  if (remaining.length > 0) {
-    throw new Error(`Unable to terminate ${remaining.length} process(es) owned by Run ${handle.runId}`);
-  }
-  await handle.exit;
-  return { signal: "SIGKILL", processAlive: false };
-};
-
-export const heartbeatRuntime = async (handle: RuntimeHandle): Promise<HeartbeatSnapshot> => {
-  const processAlive = handle.child.exitCode === null && handle.child.signalCode === null;
-  if (processAlive) handle.lastProcessAliveAt = new Date();
-  return {
-    processAlive,
-    lastProcessAliveAt: handle.lastProcessAliveAt,
-    lastProgressEventAt: handle.lastProgressEventAt,
-    inFlightTool: handle.inFlightTool,
-  };
-};
-
 export type RunnerDefinition = {
   binaryEnvironment: string;
   defaultBinary: string;
@@ -849,91 +196,42 @@ export type RunnerDefinition = {
   toolTransport: SessionToolTransport;
   toolEntrypoint(): string;
   adapter: CliAdapter;
-  /** Whether this CLI owns a per-Session config root outside disposable scratch. */
   isolatesSessionConfig: boolean;
-  /** Model identity used only for daemon startup preflight, when that adapter requires one. */
   startupPreflightModel: string | null;
   args(spec: RunSpec, resume?: ResumeSpec): string[];
   childEnvironment(claim: Pick<ClaimedTask, "run">, scratch: AgentScratch): NodeJS.ProcessEnv;
-  provisionSessionConfig(
-    config: RunnerConfig,
-    scratch: AgentScratch,
-    options?: SessionConfigOptions,
-  ): Promise<void>;
+  provisionSessionConfig(config: RunnerConfig, scratch: AgentScratch, options?: SessionConfigOptions): Promise<void>;
 };
 
-const deferredAdapter = (create: () => CliAdapter): CliAdapter => {
-  let concrete: CliAdapter | null = null;
-  const get = (): CliAdapter => {
-    concrete ??= create();
-    return concrete;
-  };
-  return Object.freeze<CliAdapter>({
-    preflight: (spec) => get().preflight(spec),
-    start: (spec, sink) => get().start(spec, sink),
-    resume: (spec, sink) => get().resume(spec, sink),
-    kill: (handle, reason) => get().kill(handle, reason),
-    heartbeat: (handle) => get().heartbeat(handle),
-    classifyError: (evidence) => get().classifyError(evidence),
-  });
-};
+const definitionFrom = (declaration: AdapterDeclaration): Readonly<RunnerDefinition> => Object.freeze({
+  binaryEnvironment: declaration.binaryEnvironment,
+  defaultBinary: declaration.defaultBinary,
+  toolIntroduction: declaration.toolIntroduction,
+  toolTransport: declaration.toolTransport,
+  toolEntrypoint: declaration.toolEntrypoint,
+  adapter: createCliAdapter(declaration),
+  isolatesSessionConfig: declaration.isolatesSessionConfig,
+  startupPreflightModel: declaration.startupPreflightModel,
+  args: declaration.args,
+  childEnvironment: declaration.childEnvironment,
+  provisionSessionConfig: declaration.provisionSessionConfig,
+});
 
-/**
- * The only per-CLI registry. Binary configuration, adapter selection,
- * availability enumeration, child config roots, and argv all derive from it.
- */
+/** The public runner registry is a derived view of provider-owned declarations. */
 export const RUNNER_DEFINITIONS: Readonly<Record<RunnerKind, Readonly<RunnerDefinition>>> = Object.freeze({
-  CLAUDE: Object.freeze<RunnerDefinition>({
-    binaryEnvironment: "CLAUDE_BINARY",
-    defaultBinary: "claude",
-    toolIntroduction: "Anneal tools attached to this session (MCP server 'agentos'; your client may prefix them, e.g. mcp__agentos__task_output):",
-    toolTransport: "mcp-stdio",
-    toolEntrypoint: mcpServerPath,
-    adapter: deferredAdapter(() => createClaudeAdapter()),
-    isolatesSessionConfig: false,
-    startupPreflightModel: null,
-    args: (spec, resume) => claudeArgs(spec, resume),
-    childEnvironment: (claim, scratch) => claudeChildEnvironment(claim, scratch),
-    provisionSessionConfig: (config, scratch, options) => provisionClaudeSessionConfig(config, scratch, options),
-  }),
-  CODEX: Object.freeze<RunnerDefinition>({
-    binaryEnvironment: "CODEX_BINARY",
-    defaultBinary: "codex",
-    toolIntroduction: "Anneal tools attached to this session (MCP server 'agentos'; your client may prefix them, e.g. mcp__agentos__task_output):",
-    toolTransport: "mcp-stdio",
-    toolEntrypoint: mcpServerPath,
-    adapter: deferredAdapter(() => createCodexAdapter()),
-    isolatesSessionConfig: true,
-    startupPreflightModel: CODEX_STARTER_MODEL,
-    args: (spec, resume) => codexArgs(spec, resume),
-    childEnvironment: (claim, scratch) => codexChildEnvironment(claim, scratch),
-    provisionSessionConfig: (config, scratch, options) => provisionCodexSessionConfig(config, scratch, options),
-  }),
-  PI: Object.freeze<RunnerDefinition>({
-    binaryEnvironment: "PI_BINARY",
-    defaultBinary: "pi",
-    toolIntroduction: "Anneal tools attached to this session (pi extension tools):",
-    toolTransport: "pi-extension",
-    toolEntrypoint: piExtensionPath,
-    adapter: deferredAdapter(() => createPiAdapter()),
-    isolatesSessionConfig: true,
-    startupPreflightModel: "openai-codex/gpt-5.6-luna",
-    args: (spec, resume) => piArgs(spec, resume),
-    childEnvironment: (claim, scratch) => piChildEnvironment(claim, scratch),
-    provisionSessionConfig: (config, scratch, options) => provisionPiSessionConfig(config, scratch, options),
-  }),
+  CLAUDE: definitionFrom(ADAPTER_DECLARATIONS.CLAUDE),
+  CODEX: definitionFrom(ADAPTER_DECLARATIONS.CODEX),
+  PI: definitionFrom(ADAPTER_DECLARATIONS.PI),
 });
 
 export const RUNNER_KINDS = Object.freeze(Object.keys(RUNNER_DEFINITIONS) as RunnerKind[]);
 
-/** Frozen: a caller substitutes an adapter by passing one to `executeClaim`,
- *  not by writing over this derived view. */
 export const adapters: Readonly<Record<RunnerKind, CliAdapter>> = Object.freeze(Object.fromEntries(
   RUNNER_KINDS.map((runner) => [runner, RUNNER_DEFINITIONS[runner].adapter]),
 ) as Record<RunnerKind, CliAdapter>);
 
-/** Hash the exact prompt bytes sent to the provider on this invocation. */
-export const promptHashFor = (prompt: string): string => createHash("sha256").update(prompt).digest("hex");
+export const argsForRunner = (runner: RunnerKind, spec: RunSpec, resume?: ResumeSpec): string[] =>
+  RUNNER_DEFINITIONS[runner].args(spec, resume);
 
 export const manifestFor = (spec: RunSpec, dispatchedPrompt: string): Record<string, unknown> => ({
   adapterVersion: ADAPTER_VERSION,
@@ -947,7 +245,6 @@ export const manifestFor = (spec: RunSpec, dispatchedPrompt: string): Record<str
   promptHash: promptHashFor(dispatchedPrompt),
   promptTransport: "stdin",
   structuredEvents: true,
-  // What the seat manual promises the agent, and how it was actually injected.
   agentosTools: {
     transport: RUNNER_DEFINITIONS[spec.claim.runner].toolTransport,
     entrypoint: RUNNER_DEFINITIONS[spec.claim.runner].toolEntrypoint(),
